move a bunch of things in waitlist and home
