+++ apps/web/src/app/_components/hero-challenge-card.tsx
