+++ apps/web/src/app/_components/hero.tsx
