+++ apps/web/src/app/waitlist/_components/waitlist-form.tsx
@@ -21,7 +21,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@repo/ui';
-import { uploadWaitlistEntry } from '~/components/landing/waitlist/create.action';
+import { uploadWaitlistEntry } from './create.action';
 
 const waitlistFormSchema = z.object({
   name: z.string().min(1, 'Please enter your name'),
