+++ apps/web/src/app/waitlist/_components/waitlist.module.css
@@ -1,0 +1,26 @@
+/* The module for the Waitlist page background */
+.waitlistBackground {
+    top: -56px;
+    background: linear-gradient(0deg, hsl(var(--background)) 0%, rgba(9, 9, 11, 0) 100%),
+      linear-gradient(
+        270deg,
+        rgba(59, 131, 199, 0) 44.51%,
+        rgba(59, 131, 199, 0.17) 75.5%,
+        rgba(49, 120, 198, 0.3) 100%
+      ),
+      linear-gradient(
+        90deg,
+        rgba(59, 199, 182, 0) 22.98%,
+        rgba(59, 182, 199, 0.17) 67.78%,
+        rgba(49, 189, 198, 0.3) 100%
+      );
+  }
+  
+  .mybg {
+    background: url('../../../assets/images/waitlist-bg.webp');
+    background-position-y: 33%;
+    background-position-x: center;
+    background-repeat: no-repeat;
+    background-size: 111% 200%;
+    filter: blur(69px);
+  }
